changed the project structure and fixed cmake (hopefully)
changed the project to have a separeate rp2040 in each fader wing for control there. so i needed 2 binaries that both used the same libraries for components. and after moving those and (what i thought) updating the cmake files everything was fucked. so about an hour and a half later and a all my gpt4o tokes for today i have fixed them